refactor: placement of test
